Consolidate tests into single app
