+++ authorizenet/tests/tests.py
@@ -1,57 +1,50 @@
 from datetime import datetime
-from django.test import TestCase
+from django.test import TestCase, LiveServerTestCase
 from xml.dom.minidom import parseString
-from httmock import urlmatch, HTTMock
+from httmock import HTTMock
 
 from authorizenet.cim import extract_form_data, extract_payment_form_data, \
     add_profile
 
-
-cim_url_match = urlmatch(scheme='https', netloc=r'^api\.authorize\.net$',
-                         path=r'^/xml/v1/request\.api$')
-
-
-success_response = (
-    '<?xml version="1.0" encoding="utf-8"?>'
-    '<{0}>'
-    '<messages>'
-    '<resultCode>Ok</resultCode>'
-    '<message><code>I00001</code><text>Successful.</text></message>'
-    '</messages>'
-    '<customerProfileId>6666</customerProfileId>'
-    '<customerPaymentProfileIdList>'
-    '<numericString>7777</numericString>'
-    '</customerPaymentProfileIdList>'
-    '<customerShippingAddressIdList />'
-    '<validationDirectResponseList />'
-    '</{0}>'
-)
+from .utils import create_user, xml_to_dict
+from .mocks import cim_url_match, success_response
 
 
-def xml_to_dict(node):
-    """Recursively convert minidom XML node to dictionary"""
-    node_data = {}
-    if node.nodeType == node.TEXT_NODE:
-        node_data = node.data
-    elif node.nodeType not in (node.DOCUMENT_NODE, node.DOCUMENT_TYPE_NODE):
-        node_data.update(node.attributes.items())
-    if node.nodeType not in (node.TEXT_NODE, node.DOCUMENT_TYPE_NODE):
-        for child in node.childNodes:
-            child_name, child_data = xml_to_dict(child)
-            if not child_data:
-                child_data = ''
-            if child_name not in node_data:
-                node_data[child_name] = child_data
-            else:
-                if not isinstance(node_data[child_name], list):
-                    node_data[child_name] = [node_data[child_name]]
-                node_data[child_name].append(child_data)
-        if node_data.keys() == ['#text']:
-            node_data = node_data['#text']
-    if node.nodeType == node.DOCUMENT_NODE:
-        return node_data
-    else:
-        return node.nodeName, node_data
+class PaymentProfileCreationTests(LiveServerTestCase):
+
+    def test_create_new_customer_get(self):
+        create_user(username='billy', password='password')
+        self.client.login(username='billy', password='password')
+        response = self.client.get('/customers/create')
+        self.assertNotIn("This field is required", response.content)
+        self.assertIn("Credit Card Number", response.content)
+        self.assertIn("City", response.content)
+
+    def test_create_new_customer_post_error(self):
+        create_user(username='billy', password='password')
+        self.client.login(username='billy', password='password')
+        response = self.client.post('/customers/create')
+        self.assertIn("This field is required", response.content)
+        self.assertIn("Credit Card Number", response.content)
+        self.assertIn("City", response.content)
+
+    def test_create_new_customer_post_success(self):
+        create_user(username='billy', password='password')
+        self.client.login(username='billy', password='password')
+        response = self.client.post('/customers/create', {
+            'card_number': "5586086832001747",
+            'expiration_date_0': "4",
+            'expiration_date_1': "2020",
+            'card_code': "123",
+            'first_name': "Billy",
+            'last_name': "Monaco",
+            'address': "101 Broadway Ave",
+            'city': "San Diego",
+            'state': "CA",
+            'country': "US",
+            'zip': "92101",
+        }, follow=True)
+        self.assertIn("success", response.content)
 
 
 class ExtractFormDataTests(TestCase):
